requirements.txt
requests
vk_api
python-qbittorrent